Fix missing container path in SCC error message when capabilities are invalid
This fixes the issue where if a pod had invalid capabilities configured,
the SCC error message did not indicate which container the error message
applies to, as is the case with other SCC error messages (e.g. runAsUser).
This was the error message:

capabilities.add: Invalid value: "NET_BIND_SERVICE": capability may not be added

This commit fixes the message. The new error message is:

spec.containers[0].securityContext.capabilities.add: Invalid value: "NET_BIND_SERVICE": capability may not be added