Fix missing "spec.securityContext" in path in SCC error message when fsGroup/supplementalGroups are invalid
This fixes the issue where if a pod had invalid fsGroup or supplementalGroups,
the SCC error message did not indicate the full field path as is the case with
other SCC error messages (e.g. spec.securityContext.seLinuxOptions).
This was the error message:

fsGroup: Invalid value: []int64{123}: 123 is not an allowed group

This commit fixes the message. The new error message is:

spec.securityContext.fsGroup: Invalid value: []int64{123}: 123 is not an allowed group