pkg/securitycontextconstraints: fix points groups
To generate a reason message we rely on the fact that point groups are not overflowing.
However, the current numbering does not respect the fact that runAs* points are counted twice,
once for RunAsUser and once for SELinuxContext constraints.

This fixes it by shifting the numbers to account for selinux runAs* point scores.